debug

+++ physcraper/wrappers.py
@@ -687,10 +687,11 @@ def concat(genelistdict, workdir_comb, email, num_threads=None, percentage=0.37,
 def save_copy_code(workdir_comb):
     i = 1
     if os.path.exists("{}/physcraper_runcopy".format(workdir_comb)):
-        while os.path.exists("{}/physcraper_runcopy{}".format(workdir_comb, i)):
-            prev_dir = "{}/physcraper_runcopy{}".format(workdir_comb, i)
-            i += 1
-            shutil.move("{}/physcraper_runcopy".format(workdir_comb), prev_dir)
+        if os.path.exists("{}/physcraper_runcopy{}".format(workdir_comb, i)):
+            while os.path.exists("{}/physcraper_runcopy{}".format(workdir_comb, i)):
+                prev_dir = "{}/physcraper_runcopy{}".format(workdir_comb, i)
+                i += 1
+                shutil.move("{}/physcraper_runcopy".format(workdir_comb), prev_dir)
     shutil.copytree("./physcraper", "{}/physcraper_runcopy".format(workdir_comb))
